Http Server: List all envinronment variable
This adds an API that returns a list of all the environment variable.

This patchs also adds a test to the testenv.py that verify that the
variable that was added in the test is part of the the list return from
the new API.

+++ modules/httpserver/tests/api/testenv.py
@@ -4,11 +4,14 @@ import basetest
 class testenv(basetest.Basetest):
     def test_env(self):
         param = "test-param"
-        get_path = self.path_by_nick(self.env_api, "getEnv") + param
+        get_path = self.path_by_nick(self.env_api, "list_env") + param
         set_path = get_path + "?val=TEST"
         self.curl(set_path, method='POST')
         val = self.curl(get_path)
         self.assertEqual(val, "TEST")
+        lst = self.curl(self.path_by_nick(self.env_api, "list_env"))
+        if not param + "=TEST" in lst:
+            raise Exception('environment variable not found in list')
 
         self.curl(get_path, method="DELETE")
         self.assertHttpError(get_path, 400)
